# Conflicts:
#	src/main/java/controller/BookDetailServlet.java
#	src/main/java/controller/BookServlet.java
#	src/main/java/controller/CartServlet.java
#	src/main/java/controller/CategoryBookServlet.java
#	src/main/java/controller/CategoryServlet.java
#	src/main/java/controller/ForgotPasswordServlet.java
#	src/main/java/controller/LoginServlet.java
#	src/main/java/controller/NotificationServlet.java
#	src/main/java/controller/OrderTrackingPageServlet.java
#	src/main/java/controller/PaymentServlet.java
#	src/main/java/controller/RegisterServlet.java
#	src/main/java/controller/ResetPasswordServlet.java
#	src/main/java/controller/ReviewServlet.java
#	src/main/java/controller/UpdateUserInfo.java
#	src/main/java/controller/UserServlet.java
#	src/main/java/controller/VerifyServlet.java
#	src/main/java/controller/ViewHistoryServlet.java
#	src/main/java/controller/WishListServlet.java
#	src/main/java/dao/AddressDao.java
#	src/main/java/dao/BookDao.java
#	src/main/java/dao/BookReviewDao.java
#	src/main/java/dao/CartDAO.java
#	src/main/java/dao/CategoryBookDao.java
#	src/main/java/dao/CategoryDao.java
#	src/main/java/dao/NotificationDao.java
#	src/main/java/dao/OrderDAO.java
#	src/main/java/dao/OrderStatusDAO.java
#	src/main/java/dao/UserDao.java
#	src/main/java/dao/ViewHistoryDao.java
#	src/main/java/dao/WishListDao.java
#	src/main/java/service/AddressService.java
#	src/main/java/service/BookReviewService.java
#	src/main/java/service/BookService.java
#	src/main/java/service/CategoryBookService.java
#	src/main/java/service/CategoryService.java
#	src/main/java/service/NotificationService.java
#	src/main/java/service/OrderService.java
#	src/main/java/service/UserService.java
#	src/main/java/service/WishListService.java
#	src/main/java/util/AuthFilter.java

+++ src/main/java/model/Review.java
@@ -21,7 +21,7 @@ public class Review {
     private Long id;
 
     @Column(name = "\"rating\"")
-    private double rating;
+    private Integer rating;
 
     @Column(name = "\"comment\"")
     private String comment;
